Attempting to address 'ReferenceError: extension not found' exception in
the gallery, which currently appears after dashboard admin login (and on root URL).
Seems to be related to a mismatch between the baseUrl value in the options object passed to routes.dashboardAdmins in server/index.js, and then the registered name of the extension. The former currently is slash-terminated (caused by the need to make the React base URL act on both slash-and-non-slash terminated root path) and the latter is not slash terminated.