Added spinner when onboarding (creating Gauss tenant). Reworked the
state mgmt around available domain so it is picked up before attempting to fix any errors relating to tenants in Gauss that are not registered in Verify.
The re-authentication is redirect-based, to the desired DNS domain must be kept in (local|session)Storage - first take uses localStorage, but should probably be changed to sessionStorage.
Refreshing the /verify page after onboarding brings back the 'domain already taken' page ATM, something is not working about getting the domain to use from the already onboarded domains.